refactor: DTO 패키지 구조를 request/response로 체계화
