+++ src/main/java/org/iebbuda/mozi/user/controller/UserController.java
@@ -3,10 +3,10 @@ package org.iebbuda.mozi.user.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import org.iebbuda.mozi.user.dto.LoginIdFindByEmailRequestDTO;
-import org.iebbuda.mozi.user.dto.LoginIdFindResponseDTO;
-import org.iebbuda.mozi.user.dto.UserJoinRequestDTO;
-import org.iebbuda.mozi.user.dto.UserJoinResponseDTO;
+import org.iebbuda.mozi.user.dto.request.LoginIdFindByEmailRequestDTO;
+import org.iebbuda.mozi.user.dto.response.LoginIdFindResponseDTO;
+import org.iebbuda.mozi.user.dto.request.UserJoinRequestDTO;
+import org.iebbuda.mozi.user.dto.response.UserJoinResponseDTO;
 import org.iebbuda.mozi.user.service.UserService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
