+++ src/main/java/org/iebbuda/mozi/user/dto/request/UserJoinRequestDTO.java
@@ -1,4 +1,4 @@
-package org.iebbuda.mozi.user.dto;
+package org.iebbuda.mozi.user.dto.request;
 
 
 import lombok.*;
