+++ src/main/java/org/iebbuda/mozi/user/service/UserServiceImpl.java
@@ -5,10 +5,10 @@ import lombok.extern.log4j.Log4j2;
 import org.iebbuda.mozi.security.account.domain.AuthVO;
 import org.iebbuda.mozi.security.account.domain.UserRole;
 import org.iebbuda.mozi.user.domain.UserVO;
-import org.iebbuda.mozi.user.dto.LoginIdFindResponseDTO;
-import org.iebbuda.mozi.user.dto.UserDTO;
-import org.iebbuda.mozi.user.dto.UserJoinResponseDTO;
-import org.iebbuda.mozi.user.dto.UserJoinRequestDTO;
+import org.iebbuda.mozi.user.dto.response.LoginIdFindResponseDTO;
+import org.iebbuda.mozi.user.dto.response.UserDTO;
+import org.iebbuda.mozi.user.dto.response.UserJoinResponseDTO;
+import org.iebbuda.mozi.user.dto.request.UserJoinRequestDTO;
 import org.iebbuda.mozi.user.mapper.UserMapper;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
